Run behat tests also on ownCloud with github actions

+++ tests/scripts/setupEnvForCi.sh
@@ -8,6 +8,12 @@
 
 # Prerequisite: The server to use is downloaded and extracted to /tmp/oc_music_ci/server
 
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 <owncloud|nextcloud>"
+    exit 1
+fi
+
+CLOUD=$1
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 REPO_DIR=$SCRIPT_DIR/../..
 
@@ -21,15 +27,19 @@ rm -rf data
 mkdir data
 cd server
 touch config/CAN_INSTALL
-php occ maintenance:install --database-name oc_autotest --database-user oc_autotest --admin-user admin --admin-pass 0aVnqOWH1rurCrNdTJTM --database sqlite --database-pass='' --data-dir=/tmp/oc_music_ci/data
+php occ maintenance:install --database-name owncloud --database-user oc_autotest --admin-user admin --admin-pass 0aVnqOWH1rurCrNdTJTM --database sqlite --database-pass='' --data-dir=/tmp/oc_music_ci/data
 OC_PASS=ampache123456 php occ user:add ampache --password-from-env
 
 # set log level as 'info'
 php occ config:system:set loglevel --type=integer --value=1
 
-# enable the Music app, removing the ownCloud-specific files first
-rm apps/music/appinfo/database.xml
-rm apps/music/appinfo/app.php
+# remove the ownCloud-specific files of Music on Nextcloud
+if [ $CLOUD == 'nextcloud' ]; then
+    rm apps/music/appinfo/database.xml
+    rm apps/music/appinfo/app.php
+fi
+
+# activate the Music app
 php occ app:enable music
 
 # download and scan the test content
@@ -38,4 +48,4 @@ php occ files:scan ampache
 php occ music:scan ampache
 
 # setup the API key
-sqlite3 /tmp/oc_music_ci/data/oc_autotest.db 'INSERT INTO oc_music_ampache_users (user_id, hash) VALUES ("ampache", "3e60b24e84cfa047e41b6867efc3239149c54696844fd3a77731d6d8bb105f18");'
+sqlite3 /tmp/oc_music_ci/data/owncloud.db 'INSERT INTO oc_music_ampache_users (user_id, hash) VALUES ("ampache", "3e60b24e84cfa047e41b6867efc3239149c54696844fd3a77731d6d8bb105f18");'
